refactor: BlogArchives widget #308
