+++ src/Widgets/BlogArchives/Components/BlogArchives.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Fan.Blog.Models
-@using Fan.WebApp.Widgets.BlogArchives
+@using BlogArchives
 @model Tuple<Dictionary<int, List<MonthItem>>, BlogArchivesWidget>
 @{
     var years = Model.Item1;
