+++ src/Widgets/BlogArchives/Manage/Widgets/BlogArchivesEdit.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model BlogArchivesEditModel
+@model BlogArchives.Manage.Widgets.BlogArchivesEditModel
 @{
     ViewData["Title"] = "BlogArchives Widget Settings";
     Layout = "_SettingsLayout";
